[Fixed] Bolma has meat now [Updated] Fireworks [Updated] Skill Mods [Added] Skill mod bonuses display [Added] skillMods vector to commands [Added] Redeed conditions for houses, harvesters and factories. [Fixed] Custom names on factory crates [Added] Error when factory output hopper is full [Fixed] Power message when out of maint [Updated] Tripled trap xp [Removed] Random jedi skill mod code
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@4592 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/commands/ThrowTrapCommand.h
@@ -187,7 +187,7 @@ public:
 				if(state != 0)
 					buff->addState(state);
 
-				VectorMap<String, int64>* skillMods = trapData->getSkillMods();
+				VectorMap<String, int>* skillMods = trapData->getSkillMods();
 				for(int i = 0; i < skillMods->size(); ++i) {
 					buff->setSkillModifier(skillMods->elementAt(i).getKey(), skillMods->get(i));
 				}
